Cart con objeto ApiResponse + rutas personalizadas + Authorize en todas las rutas + manejo de errores

+++ CarritoDeCompras/Controllers/CartController.cs
@@ -6,6 +6,8 @@ using Microsoft.AspNetCore.Mvc;
 using Capa.Aplicacion.DTI;
 using Microsoft.AspNetCore.Authorization;
 using Capa.Infraestructura.Servicios.Utilidades;
+using Capa.Datos.Modelos;
+using Azure;
 
 
 // For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
@@ -27,60 +29,97 @@ namespace CarritoDeCompras.Controllers
             _httpAccesor = httpAccesor;
         }
 
-        [HttpGet]
+        [HttpGet("getCart")]
         [Authorize]
         public async Task<ActionResult<CartDTO>> Get()
         {
+            ApiResponse<CartDTO> response;
+
             try
             {
                 var userId = _httpAccesor.getUserIdToken();
+
+                if (userId == null)
+                {
+                    response = ApiResponse<CartDTO>.ErrorResponse(404, "Usuario no encontrado");
+                    return NotFound(response);
+                }
+
                 var cart = await _cartService.GetCartById(userId);
 
                 if (cart != null)
                 {
                     var cartDto = _mapper.Map<CartDTO>(cart);
-                    return Ok(cartDto);
+                    response = ApiResponse<CartDTO>.SuccessResponse(cartDto, 200);
+                    return Ok(response);
                 }
 
-                return NotFound("Carrito no encontrado!");
+                response = ApiResponse<CartDTO>.ErrorResponse(404, "El usuario actual no tiene un carrito asociado");
+                return NotFound(response);
             }
             catch (Exception ex)
             {
-                return BadRequest(ex.Message);
+                response = ApiResponse<CartDTO>.ErrorResponse(500, ex.Message);
+                return BadRequest(response);
             }
 
         }
 
-        [HttpPost("Crear")]
+        [HttpPost("newCart")]
         [Authorize]
         public async Task<IActionResult> Create()
         {
+            ApiResponse<Cart> response;
+
             try
             {
                 var userId = _httpAccesor.getUserIdToken();
+
+                if (userId == null)
+                {
+                    response = ApiResponse<Cart>.ErrorResponse(404, "Usuario no encontrado");
+                    return NotFound(response);
+                }
 
                 var cart = new Cart();
                 cart.UserId = userId;
 
                 await _cartService.createCart(cart);
 
-                return Ok("Carrito creado correctamente");
+                response = ApiResponse<Cart>.SuccessResponse(cart, 201, "Carrito creado correctamente!");
+                return Ok(response);
             }
 
             catch (Exception ex)
             {
-                return BadRequest(ex.Message);
+                response = ApiResponse<Cart>.ErrorResponse(500, ex.Message);
+                return BadRequest(response);
             }
         }
 
-        [HttpPost("AddProduct")]
+        [HttpPost("addProduct")]
         [Authorize]
         public async Task<IActionResult> Add([FromBody] CartItemDTI cartItem)
         {
+            ApiResponse<CartItemDTI> response;
+
             try
             {
                 var userId = _httpAccesor.getUserIdToken();
+
+                if (userId == null)
+                {
+                    response = ApiResponse<CartItemDTI>.ErrorResponse(404, "Usuario no encontrado");
+                    return NotFound(response);
+                }
+
                 var cart = await _cartService.GetCartById(userId);
+
+                if (cart == null)
+                {
+                    response = ApiResponse<CartItemDTI>.ErrorResponse(404, "Carrito no encontrado!");
+                    return NotFound(response);
+                }
 
                 var item = new CartItem
                 {
@@ -89,36 +128,79 @@ namespace CarritoDeCompras.Controllers
                     Cantidad = cartItem.Cantidad,
                 };
 
+                var itemDti = _mapper.Map<CartItemDTI>(item);
+
                 await _cartService.AddProduct(item, userId);
 
-                return Ok("Producto agregado al carrito correctamente");
+                response = ApiResponse<CartItemDTI>.SuccessResponse(itemDti, 200, "Producto agregado al carrito correctamente");
+                return Ok(response);
 
             }
             catch (Exception ex)
             {
-                return BadRequest();
+                response = ApiResponse<CartItemDTI>.ErrorResponse(500, ex.Message);
+                return BadRequest(response);
             }
 
         }
 
-        [HttpDelete]
-        public async Task<IActionResult> Delete(int cartId, [FromBody] CartItemDTI cartItemDTI)
+        [HttpPost("resetCart")]
+        [Authorize]
+        public async Task<IActionResult> ResetCart()
         {
+            ApiResponse<string> response;
+
             try
             {
                 var userId = _httpAccesor.getUserIdToken();
+
+                if (userId == null)
+                {
+                    response = ApiResponse<string>.ErrorResponse(404, "Usuario no encontrado");
+                    return NotFound(response);
+                }
+
+                await _cartService.ResetCart(userId);
+                response = ApiResponse<string>.SuccessResponse("Se vació el carrito", 200);
+                return Ok(response);
+
+            }
+            catch (Exception ex)
+            {
+                response = ApiResponse<string>.ErrorResponse(500, ex.Message);
+                return BadRequest(response);
+            }
+        }
+
+        [HttpDelete("deleteProduct")]
+        [Authorize]
+        public async Task<IActionResult> Delete(int cartId, [FromBody] CartItemDTI cartItemDTI)
+        {
+            ApiResponse<CartItemDTI> response;
+
+            try
+            {
+                var userId = _httpAccesor.getUserIdToken();
+
+                if (userId == null)
+                {
+                    response = ApiResponse<CartItemDTI>.ErrorResponse(404, "Usuario no encontrado");
+                    return NotFound(response);
+                }
 
                 var cartItem = _mapper.Map<CartItem>(cartItemDTI);
 
                 await _cartService.RemoveProduct(cartItem, userId);
 
-                return Ok("Producto elimado del carrito correctamente");
+                response = ApiResponse<CartItemDTI>.SuccessResponse(cartItemDTI, 200, "Producto eliminado al carrito correctamente");
+
+                return Ok(response);
             }
             catch (Exception ex)
             {
-                return BadRequest();
+                response = ApiResponse<CartItemDTI>.ErrorResponse(500, ex.Message);
+                return BadRequest(response);
             }
-
         }
     }
 }
